+++ package.json
@@ -8,8 +8,9 @@
   "repository": "https://github.com/LambdaSchool/webdb-i-challenge",
   "dependencies": {
     "express": "^4.17.1",
-    "knex": "^0.20.1",
-    "sqlite3": "^4.1.0"
+    "express-async-handler": "^1.1.4",
+    "knex": "^0.20.8",
+    "sqlite3": "^4.1.1"
   },
   "devDependencies": {
     "nodemon": "^1.19.4"
